+++ src/Components/Favorites.tsx
@@ -1,0 +1,14 @@
+import React, { useEffect, useState } from 'react';
+
+// import ResultsPanel from './SearchPanel';
+import { Header, Button } from './defaultComponents';
+
+function Favorites() {  
+  return (
+    <div>
+      <Button>Go Home</Button>
+    </div>
+  );
+}
+
+export default Favorites;
