+++ src/Components/View.tsx
@@ -1,0 +1,14 @@
+import React, { useEffect, useState } from 'react';
+
+// import ResultsPanel from './SearchPanel';
+// import { Header, Button } from './defaultComponents';
+
+function View() {  
+  return (
+    <div>
+      view
+    </div>
+  );
+}
+
+export default View;
