+++ src/Components/defaultComponents.tsx
@@ -46,4 +46,53 @@ const Button = styled.button<any>`
 `;
 
 
-export { Header, Button, HotSearchInput }
+const Panel = styled.div`
+  display: grid;
+  text-align: center;
+  
+  @media screen and (max-width: 480px){
+    grid-template-columns: repeat(1, 1fr);
+  }
+
+  @media screen and (min-width: 480px){
+    grid-template-columns: repeat(2, 1fr);
+  }
+  
+  @media screen and (min-width: 768px){
+    grid-template-columns: repeat(4, 1fr);
+  }
+
+  @media screen and (min-width: 1024px){
+    grid-template-columns: repeat(5, 1fr);
+  }
+
+  @media screen and (min-width: 1920px){
+    grid-template-columns: repeat(10, 1fr);
+  }
+`;
+
+const Card = styled.div`
+  padding: 10px;
+  
+  background: white;
+  border-radius: 10px;
+  box-shadow: 0px 0px 2px 1px #8080809f;
+  text-align: center;
+
+  &:hover {
+    background: bisque;
+    cursor: pointer;
+  }
+`;
+
+const Card__name = styled.div`
+  padding: 2px 0;
+  overflow-wrap: anywhere;
+`;
+
+const Avatar = styled.img`
+  width: 100px;
+  height: 100px;
+`;
+
+export { Header, Button, HotSearchInput, Panel, Card, Card__name, Avatar }
